requirements.txt
requests==2.31.0
pandas==2.0.3
pymongo==4.5.0
python-dotenv==1.0.0
schedule==1.2.0
numpy==1.24.3
python-telegram-bot==20.7
